refactor client base
