+++ src/talemate/client/__init__.py
@@ -1,3 +1,4 @@
+import os
 from talemate.client.openai import OpenAIClient
 from talemate.client.registry import CLIENT_CLASSES, get_client_class, register
 from talemate.client.textgenwebui import TextGeneratorWebuiClient
