+++ src/talemate/server/run.py
@@ -1,3 +1,5 @@
+import os
+
 import argparse
 import asyncio
 import sys
